backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pyjwt
passlib
bcrypt
